[Console] Move QuestionPrompt classes to Console\Question namespace
